create express file to seperate it from the server logic
